# 8256523: Streamline Java SHA2 implementation